Turning off clustering.
Trying to avoid the dreaded "duplicate key" error when conqueso starts.

The theory is that there is a race condition with the migrations (you can simulate this by turning off the migrations and restarting).  The clustering was originally added to help with performance but the performance is so good now, it's not needed.  @swillard is working on a revamp of the code to yank it out officially.